- adjusted the payment form design - changed vat from float to decimal

+++ DAL/OrderItemDao.cs
@@ -46,7 +46,7 @@ namespace DAL
                     Name = row["name"].ToString(),
                     Item_Category = row["item_category"].ToString(),
                     Stock = (int)row["stock"],
-                    Vat = Convert.ToDouble(row["vat"]),
+                    Vat = Convert.ToDecimal(row["vat"]),
                     Price = (decimal)row["price"],
                     Preperation_Time = (int)row["preparation_time"],
                     Menu_Id = (int)row["menu_id"]
